Asynchronous invocations.
The idea is that we will wrap libsvn_client calls in these asynchronous
calls, so SVN ops will run in auxiliary threads.

TODO: make a thread pool of some kind? Is spawning threads cheap enough?

Also, we will probably have an authentication provider that wraps keychain
services. Start on that.